Rename `argdefFilename` to `semanticsFilename`.
