+++ cpgqueryingtests/src/test/resources/default.semantics
